pyproject.toml
[tool.poetry]
name = "dal"
version = "0.1.16"
description = ""
authors = ["roman-dector <roman.schaposhnikov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
peewee = "^3.15.4"
aiohttp = "^3.8.4"
aiofiles = "^23.1.0"
aiosqlite = "^0.19.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
ionex = { git = "git@github.com:gnss-lab/ionex.git", branch = "master" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
